Cleanup cert registration form and associated tested
